[Fixes #7842] Rename layer to dataset - fixing assorted issues with default property values
